requirements.txt
agno
python-dotenv
aiohttp
openai
pgvector
psycopg2-binary
sqlalchemy
uvicorn
aiofiles
tiktoken
chonkie